pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-support-chatbot"
version = "0.1.0"
description = "LangGraph-based chatbot for API support using Azure OpenAI and MCP tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Dmitry Verbichenko", email = "dmitry.verbichenko@example.com" },
]
keywords = [
    "chatbot",
    "langgraph",
    "mcp",
    "azure-openai",
    "api-support"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.urls]
Homepage = "https://github.com/verbichenko/api-support-chatbot"
Repository = "https://github.com/verbichenko/api-support-chatbot.git"
Issues = "https://github.com/verbichenko/api-support-chatbot/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/api_support_chatbot"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "N806", "N803"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true